+++ chain/x/oracle/keeper/querier.go
@@ -95,10 +95,19 @@ func queryRequestByID(ctx sdk.Context, path []string, k Keeper) ([]byte, error) 
 		return nil, err
 	}
 	reports := k.GetReports(ctx, types.RequestID(id))
-	result, _ := k.GetResult(ctx, types.RequestID(id))
+
+	if !k.HasResult(ctx, types.RequestID(id)) {
+		return codec.MarshalJSONIndent(types.ModuleCdc, types.QueryRequestResult{
+			Request: request,
+			Reports: reports,
+			Result:  nil,
+		})
+	}
+
+	result := k.MustGetResult(ctx, types.RequestID(id))
 	return codec.MarshalJSONIndent(types.ModuleCdc, types.QueryRequestResult{
 		Request: request,
 		Reports: reports,
-		Result:  result,
+		Result:  &result,
 	})
 }