+++ chain/x/oracle/keeper/result.go
@@ -4,6 +4,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
+	"github.com/bandprotocol/bandchain/chain/pkg/obi"
 	"github.com/bandprotocol/bandchain/chain/x/oracle/types"
 )
 
@@ -18,13 +19,28 @@ func (k Keeper) SetResult(ctx sdk.Context, reqID types.RequestID, result []byte)
 	store.Set(types.ResultStoreKey(reqID), result)
 }
 
-// GetResult returns the result bytes for the given request ID or error if not exists.
-func (k Keeper) GetResult(ctx sdk.Context, id types.RequestID) ([]byte, error) {
+// GetResult returns the result for the given request ID or error if not exists.
+func (k Keeper) GetResult(ctx sdk.Context, id types.RequestID) (types.Result, error) {
 	bz := ctx.KVStore(k.storeKey).Get(types.ResultStoreKey(id))
 	if bz == nil {
-		return nil, sdkerrors.Wrapf(types.ErrResultNotFound, "id: %d", id)
+		return types.Result{}, sdkerrors.Wrapf(types.ErrResultNotFound, "id: %d", id)
 	}
-	return bz, nil
+	var result types.Result
+	err := obi.Decode(bz, &result)
+	if err != nil {
+		return types.Result{}, types.ErrOBIDecode
+	}
+
+	return result, nil
+}
+
+// MustGetResult returns the result for the given request ID. Panics on error.
+func (k Keeper) MustGetResult(ctx sdk.Context, id types.RequestID) types.Result {
+	result, err := k.GetResult(ctx, id)
+	if err != nil {
+		panic(err)
+	}
+	return result
 }
 
 // GetAllResults returns the list of all results in the store. Nil will be added for skipped results.