+++ chain/x/oracle/keeper/result_test.go
@@ -16,12 +16,19 @@ func TestResultBasicFunctions(t *testing.T) {
 	k.SetResult(ctx, types.RequestID(1), encodedResult)
 
 	// Test GetResult func
-	encodedResultReqID1, err := k.GetResult(ctx, types.RequestID(1))
+	result, err := k.GetResult(ctx, types.RequestID(1))
 	require.NoError(t, err)
-	require.Equal(t, encodedResult, encodedResultReqID1)
+	require.Equal(t, result, types.Result{RequestPacketData: req, ResponsePacketData: res})
+
+	// Test MustGetResult func
+	result = k.MustGetResult(ctx, types.RequestID(1))
+	require.NoError(t, err)
+	require.Equal(t, result, types.Result{RequestPacketData: req, ResponsePacketData: res})
 
 	_, err = k.GetResult(ctx, types.RequestID(2))
 	require.Error(t, err)
+
+	require.Panics(t, func() { k.MustGetResult(ctx, types.RequestID(2)) })
 
 	// Test HasResult func
 	require.True(t, k.HasResult(ctx, types.RequestID(1)))